requirements.txt
fastapi[standard]
prisma
pydantic
python-dotenv
passlib
bcrypt==4.3.0
python-jose
cryptography
redis
google-cloud-storage
requests
resend
tenacity
redis

sentence-transformers
transformers
torch
pillow
opencv-python
google-cloud-vision
rake-nltk
nltk
vaderSentiment
language-tool-python
pyyaml
numpy
tqdm
pdfplumber